+++ packages/types/src/phase0/index.ts
@@ -1,2 +1,2 @@
 export * from "./types";
-export * from "./ssz";
+export * from "./sszTypes";
