+++ packages/types/src/phase0/sszTypes.ts
@@ -1,20 +1,106 @@
-import {IPhase0Params} from "@chainsafe/lodestar-params";
-import {BitListType, BitVectorType, ContainerType, List, ListType, RootType, Vector, VectorType} from "@chainsafe/ssz";
-import {IPrimitiveSSZTypes} from "../../primitive/IPrimitiveSSZTypes";
-import {StringType} from "../../primitive/string";
-import {LazyVariable} from "../../utils/lazyVar";
-import {ValidatorStatus} from "../types";
-import * as phase0 from "../types";
 import {
+  IPhase0Params,
   ATTESTATION_SUBNET_COUNT,
   DEPOSIT_CONTRACT_TREE_DEPTH,
   JUSTIFICATION_BITS_LENGTH,
   MAX_REQUEST_BLOCKS,
   P2P_ERROR_MESSAGE_MAX_LENGTH,
-} from "./constants";
-import {IPhase0SSZTypesOnly} from "./interface";
+} from "@chainsafe/lodestar-params";
+import {
+  BigIntUintType,
+  BitListType,
+  BitVectorType,
+  ContainerType,
+  List,
+  ListType,
+  RootType,
+  Vector,
+  VectorType,
+} from "@chainsafe/ssz";
+import {PrimitiveSSZTypes} from "../primitive";
+import {StringType} from "../utils/StringType";
+import {LazyVariable} from "../utils/lazyVar";
+import {ValidatorStatus} from "./types";
+import * as phase0 from "./types";
 
-export function getPhase0Types(params: IPhase0Params, primitive: IPrimitiveSSZTypes): IPhase0SSZTypesOnly {
+export type Phase0SSZTypes = {
+  // misc
+  AttestationSubnets: BitVectorType;
+  BeaconBlockHeader: ContainerType<phase0.BeaconBlockHeader>;
+  SignedBeaconBlockHeader: ContainerType<phase0.SignedBeaconBlockHeader>;
+  Checkpoint: ContainerType<phase0.Checkpoint>;
+  CommitteeBits: BitListType;
+  CommitteeIndices: ListType<List<phase0.ValidatorIndex>>;
+  DepositMessage: ContainerType<phase0.DepositMessage>;
+  DepositData: ContainerType<phase0.DepositData>;
+  DepositDataRootList: ListType<List<phase0.Root>>;
+  DepositEvent: ContainerType<phase0.DepositEvent>;
+  Eth1Data: ContainerType<phase0.Eth1Data>;
+  Eth1DataOrdered: ContainerType<phase0.Eth1DataOrdered>;
+  Fork: ContainerType<phase0.Fork>;
+  ForkData: ContainerType<phase0.ForkData>;
+  ENRForkID: ContainerType<phase0.ENRForkID>;
+  HistoricalBlockRoots: VectorType<Vector<phase0.Root>>;
+  HistoricalStateRoots: VectorType<Vector<phase0.Root>>;
+  HistoricalBatch: ContainerType<phase0.HistoricalBatch>;
+  SlotRoot: ContainerType<phase0.SlotRoot>;
+  Validator: ContainerType<phase0.Validator>;
+  AttestationData: ContainerType<phase0.AttestationData>;
+  IndexedAttestation: ContainerType<phase0.IndexedAttestation>;
+  PendingAttestation: ContainerType<phase0.PendingAttestation>;
+  SigningData: ContainerType<phase0.SigningData>;
+  // operations
+  Attestation: ContainerType<phase0.Attestation>;
+  AttesterSlashing: ContainerType<phase0.AttesterSlashing>;
+  Deposit: ContainerType<phase0.Deposit>;
+  ProposerSlashing: ContainerType<phase0.ProposerSlashing>;
+  VoluntaryExit: ContainerType<phase0.VoluntaryExit>;
+  SignedVoluntaryExit: ContainerType<phase0.SignedVoluntaryExit>;
+  // block
+  BeaconBlockBody: ContainerType<phase0.BeaconBlockBody>;
+  BeaconBlock: ContainerType<phase0.BeaconBlock>;
+  SignedBeaconBlock: ContainerType<phase0.SignedBeaconBlock>;
+  // state
+  EpochAttestations: ListType<List<phase0.PendingAttestation>>;
+  BeaconState: ContainerType<phase0.BeaconState>;
+  // Validator
+  CommitteeAssignment: ContainerType<phase0.CommitteeAssignment>;
+  AggregateAndProof: ContainerType<phase0.AggregateAndProof>;
+  SignedAggregateAndProof: ContainerType<phase0.SignedAggregateAndProof>;
+  // ReqResp
+  Status: ContainerType<phase0.Status>;
+  Goodbye: BigIntUintType;
+  Ping: BigIntUintType;
+  Metadata: ContainerType<phase0.Metadata>;
+  BeaconBlocksByRangeRequest: ContainerType<phase0.BeaconBlocksByRangeRequest>;
+  BeaconBlocksByRootRequest: ListType<List<phase0.BeaconBlocksByRootRequest>>;
+  P2pErrorMessage: ListType<phase0.P2pErrorMessage>;
+  //api
+  AttesterDuty: ContainerType<phase0.AttesterDuty>;
+  AttesterDutiesApi: ContainerType<phase0.AttesterDutiesApi>;
+  BeaconCommitteeResponse: ContainerType<phase0.BeaconCommitteeResponse>;
+  BeaconCommitteeSubscription: ContainerType<phase0.BeaconCommitteeSubscription>;
+  BlockEventPayload: ContainerType<phase0.BlockEventPayload>;
+  ChainHead: ContainerType<phase0.ChainHead>;
+  ChainReorg: ContainerType<phase0.ChainReorg>;
+  Contract: ContainerType<phase0.Contract>;
+  FinalityCheckpoints: ContainerType<phase0.FinalityCheckpoints>;
+  FinalizedCheckpoint: ContainerType<phase0.FinalizedCheckpoint>;
+  Genesis: ContainerType<phase0.Genesis>;
+  ProposerDuty: ContainerType<phase0.ProposerDuty>;
+  ProposerDutiesApi: ContainerType<phase0.ProposerDutiesApi>;
+  SignedBeaconHeaderResponse: ContainerType<phase0.SignedBeaconHeaderResponse>;
+  SubscribeToCommitteeSubnetPayload: ContainerType<phase0.SubscribeToCommitteeSubnetPayload>;
+  SyncingStatus: ContainerType<phase0.SyncingStatus>;
+  ValidatorBalance: ContainerType<phase0.ValidatorBalance>;
+  ValidatorResponse: ContainerType<phase0.ValidatorResponse>;
+  // Non-speced types
+  SlashingProtectionBlock: ContainerType<phase0.SlashingProtectionBlock>;
+  SlashingProtectionAttestation: ContainerType<phase0.SlashingProtectionAttestation>;
+  SlashingProtectionAttestationLowerBound: ContainerType<phase0.SlashingProtectionAttestationLowerBound>;
+};
+
+export function getPhase0Types(params: IPhase0Params, primitive: PrimitiveSSZTypes): Phase0SSZTypes {
   const {
     BLSSignature,
     BLSPubkey,
@@ -35,7 +121,10 @@ export function getPhase0Types(params: IPhase0Params, primitive: IPrimitiveSSZTy
   } = primitive;
 
   // So the expandedRoots can be referenced, and break the circular dependency
-  const typesRef = new LazyVariable<IPhase0SSZTypesOnly>();
+  const typesRef = new LazyVariable<{
+    BeaconBlock: ContainerType<phase0.BeaconBlock>;
+    BeaconState: ContainerType<phase0.BeaconState>;
+  }>();
 
   // Misc types
   // ==========
@@ -588,7 +677,9 @@ export function getPhase0Types(params: IPhase0Params, primitive: IPrimitiveSSZTy
     },
   });
 
-  const phase0Types: IPhase0SSZTypesOnly = {
+  typesRef.set({BeaconBlock, BeaconState});
+
+  return {
     // misc
     Fork,
     ForkData,
@@ -666,8 +757,4 @@ export function getPhase0Types(params: IPhase0Params, primitive: IPrimitiveSSZTy
     SlashingProtectionAttestation,
     SlashingProtectionAttestationLowerBound,
   };
-
-  typesRef.set(phase0Types);
-
-  return phase0Types;
 }
