+++ packages/types/src/presets/minimal.ts
@@ -1,8 +1,4 @@
-/**
- * @module sszTypes/presets/minimal
- */
 import {params} from "@chainsafe/lodestar-params/minimal";
+import {createIBeaconSSZTypes} from "../index";
 
-import {createIBeaconSSZTypes, IBeaconSSZTypes} from "../IBeaconSSZTypes";
-
-export const types: IBeaconSSZTypes = createIBeaconSSZTypes(params);
+export const types = createIBeaconSSZTypes(params);
