+++ packages/types/src/primitive/types.ts
@@ -1,7 +1,3 @@
-/**
- * @module types
- */
-
 // Each type exported here contains both a compile-time type
 // (a typescript interface) and a run-time ssz type (a javascript variable)
 // For more information, see ./index.ts
