+++ packages/types/src/utils/StringType.ts
